refactor: split EventChart into focused sub-components
